添加Json序列化数组和反序列化数组的单元测试。

+++ src/test/java/com/beinet/firstpg/serializeTest/SerializeDto.java
@@ -12,7 +12,7 @@ public class SerializeDto {
     private String name;
     @JsonFormat(locale = "zh", timezone = "GMT+8", pattern = "yyyy-MM-dd HH:mm")
     private LocalDateTime birthday;
-    private float money;
+    private double money;
 
     @Override
     public String toString() {
